Update user model and migrations to match Laravel 6.0 defaults

+++ src/migrations/2015_02_07_172633_create_role_user_table.php
@@ -26,7 +26,7 @@ class CreateRoleUserTable extends Migration
             $table->increments('id');
 
             $table->integer('role_id')->unsigned()->index()->foreign()->references("id")->on("roles")->onDelete("cascade");
-            $table->integer('user_id')->unsigned()->index()->foreign()->references("id")->on("users")->onDelete("cascade");
+            $table->bigInteger('user_id')->unsigned()->index()->foreign()->references("id")->on("users")->onDelete("cascade");
 
             $table->timestamps();
